backend/chat_function/requirements.txt
azure-functions
azure-storage-blob
openai
azure-search-documents 